Resolve current offset when batch consists of only messages from before the requested offset
Apparently this can happen when you get compacted topics, which can
cause you to get stuck, because there's no way to resolve any offsets.
What we do now in that case is to resolve the fetched offset, which will
advance by one.

We can't just use the highwatermark because you can get a batch
consisting only of these "old" compacted records while still having lots
of valid messages afterwards (so you don't want to skip to the end of
the partition).

Hopefully fixes #562