Delete the entry for the waiter when the timeout is reached
When the lock times out the promise rejects, and the calling code will do "something sensible". Something sensible
could be to retry later -- but that actually requires the lock to be able to be acquired again. Right now this is
not possible, because the failed `tryToAcquire` is still inside the waiting queue -- and will never release the lock.