Make Array shuffle test pass in Node >= 11.0.0
Due to a change in how array sorting works, the shuffle test would fail in Node >= 11.0.0.
Instead of relying on implementation details of how the engine implements sorting, we'll
add a requirement on the shuffle implementation to always return a different order than
the input array. That way we can simply assert that the returned array is different
from the input, rather than asserting on the exact order, which may differ between engine versions.