Remove unneeded hardcoded `node_modules/.bin` references
The PATH environment variable will contain the $(npm bin) directory when running package.json scripts, and that
might be different from `node_modules/.bin` depending on the environment.

For `test:debug` we need a path (we're running `node` here, which then calls the real jest), so use `$(npm bin)`
for safety.