Don't bother applying a regular expression to what should be a number
KafkaJS keeps track of offsets as strings and Longs to avoid problems with 64bit offsets. This
does require frequent string->Long and Long->string conversions so that the Long parts don't leak
into the interface.

In the case of offsets in the offset manager we might also have undefined/null or empty offset
strings, and isInvalidOffset needs to be able to detect these. But: We don't expect to have arbitrary
strings here, and figuring out whether a thing is undefined/null/'' doesn't need a regular expression
that checks the complete string -- a simple '!offset' is just fine.