Implement generic support for client-side throttling
When a decoded response contains a `clientSideThrottleTime` property, then we understand
this to mean "do not send further requests for this number of ms on this connection".

For compatibility this is an _additional_ property, and responses in APIs that support
the client-side throttling should still continue to include a `throttleTime` field with a
value of 0 to indicate that the server did not do any throttling.

See https://cwiki.apache.org/confluence/display/KAFKA/KIP-219+-+Improve+quota+communication