Fix the requestQueue.fulfillRequest test for when there are pending requests
The test was essentially testing its own setup: The outer `beforeEach` would put the `request`
into the queue, which means the `send` method would get called (no pending/inflight requests
at that point). Then the inner `beforeEach` would saturate the queue, and then the test would
add the same request again (this time queued as pending).

When fulfilling the request with this correlation id then we would complete the first request, and
then pop the request added by the test into "in-flight" mode: At this point this.inflight[correlationId]
actually points to the second request, because `send` will have overwritten the entry.

fulfillRequest then does `this.inflight.delete(correlationId)` (assuming to delete the just-fulfilled
entry), but effectively clears out the inflight-ness of the second request.

This commit fixes the problem in the test:
1. Explicitly push the request into the queue in the tests, not in the `beforeEach` call, and
2. Pick one of the dummies created by the inner `beforeEach` to fulfill, and
3. Fix the inflight size check: We had N, and this test removes one and then checks that another
   takes its place -- so the expected size is the same size as before.

Note that this also points to another (theoretical) problem here: If a request is arriving with
the same correlation id as another request, and both of these end up being "in-flight" at the same
time, we would notice the conflict.