Implement support for follower-fetching (KIP-392)
This commit makes KafkaJS understand the "preferred replica" response that Kafka
2.4+ will include in fetch v11 responses, and use that preferred replica rather than the leader when
fetching.
The information for the preferred replica is retained for up to the "metadata max age" configured for
the client, if no update has been received until then we reset back to fetching from the leader.

The implementation is strongly inspired by the changes for the Java client[1,2], and based on initial
work in https://github.com/figma/kafkajs/pull/1 by @jgao54.

[1] https://issues.apache.org/jira/browse/KAFKA-8365
[2] https://github.com/apache/kafka/pull/6731