Wait for consumer to join the group before testing behavior
When `consumer.run()` returns it doesn't mean that the consumer is now part of the group, but
rather that we at least once tried to connect and sync. Due to errors, delays, etc the actual
joining might only happen later.

This likely fixes one of the cases where the tests "import after the jest environment is torn
down". Note that this pattern of "don't await run, and then await waitForConsumerToJoinGroup"
is used in other tests, but there also also plenty of tests using the simpler "await run" approach
that is also documented as the canonical way of handling things.