Avoid repeated costly copies of buffers when working with encoders
Instead of using `Buffer.concat` whenever writing into the encoder, and thereby copying the byte
contents of earlier added content potentially many times, this commit changes the encoder to use
a bigger pre-allocated buffer to write into directly. If that buffer fills up it a new one will be created
and the old contents copied into that.