Add new protocol errors
None of these new errors have any impact that would currently require code changes, as they
are all in relation to functionality that KafkaJS doesn't currently have.

The one exception being INVALID_RECORD, which we currently could get when producing, in which
case we would just log that it's an unknown protocol error. When Produce v8 is implemented,
this error could be improved by including more information about the invalid record.