Ensure fair fetch response allocation across topic-partitions
When receiving a fetch request, the broker processes partitions in
the order they appear in the fetch request. The response will contain
up to `maxBytesPerPartition` per partition, up to a total of `maxBytes`,
so if `maxBytesPerPartition * partitions > maxBytes`, the partitions
that appear later in the request will not be allocated any data, which
means that the consumer would first consume until it hits the head
of the first partitions, before consuming any messages on the partitions
that appear later in the request.

To ensure a more balanced consumption across all partitions, we
randomize the order of all partitions in the request. This means that
over time all partitions should be consumed approximately evenly.

Fixes #849