Fix record batch compression masking
Fixes #911

We were accidentally using the masking value for the Message type rather than
the RecordBatch type (3 instead of 7). This worked for other compression
types than zstd, but failed for zstd.

The reason is because to get the last 3 bytes of an int16, we need to AND it with 7, not 3. But for certain values, it will work anyway. Using Snappy as an example. Snappy = 2 (010 in binary). The compression masking value for the Message type is 3 (011 in binary). 2 & 3 = 2.

010
011
===
010

The correct masking value = 7. 2 & 7 = 2.

010
111
===
010

Now Zstd is 4 (100 in binary). 4 & 3 = 0.

100
011
===
000

The correct masking value = 7. 4 & 7 = 4.

100
111
===
000