Improve the available type information
This moves JSDoc comments to the correct place, fixes them to properly describe the
provided option objects, and generally aligns the types in comments and index.d.ts
with reality. If "obvious", then nested types in index.d.ts have been extended to
give them names and use them in more places.

Note that this also fixes the imports of the `EventEmitter` class in all places to
fix an error reported by VSCode.