Handle partition reassignment in fetcher
If a partition is reassigned to a different broker, it's
possible that a fetcher currently has a batch for that
topic-partition in the worker queue. This would cause double
processing. To avoid this, the fetch manager now keeps track
of which fetcher is currently handling each topic-partition,
and the fetchers will filter out batches for any topic-partition
that is currently being processed by a different fetcher.