Increase default authentication timeout
Brings it more in line with the Java client and librdkafka. The configuration
values for those clients are a bit different, where a single configuration
value is used as a "total timeout for socket establishment", instead
of having one configuration value for TCP socket connection and one
for authentication. KIP-601 goes into this in more detail.

This change does not make KafkaJS work the same way, but it does set the
default to a more sane value.