Fix invalid KafkaMessage type
The `size` attribute is only ever present on pre-0.10 `Messages`,
and never on the new `Record`. The field was mistakenly defined as
being non-optional.

Providing a meaningful `size` field for Record is tricky because in
RecordBatch all the records are compressed together, rather than individually
as is the case with the old Message protocol. Therefore you can only
calculate a size for the uncompressed record, which isn't very useful
since you most likely care about the size because you want to understand
the size over the network.